[BAEL-7669] Switched to spring-boot-application to include more examples
* Added things needed in pom.xml, without affecting existing versions
