--- spring-core-4/src/test/java/com/baeldung/globaltimezone/GlobalTimeZoneBean.java
@@ -1,17 +1,0 @@
-package com.baeldung.globaltimezone;
-
-import java.util.TimeZone;
-
-public class GlobalTimeZoneBean {
-
-    private final String globalTimeZone;
-
-    public GlobalTimeZoneBean() {
-        this.globalTimeZone = TimeZone.getDefault()
-            .getDisplayName();
-    }
-
-    public String getGlobalTimeZone() {
-        return globalTimeZone;
-    }
-}
