+++ spring-core-4/src/test/java/com/baeldung/globaltimezone/GlobalTimezoneBeanFactoryPostProcessorUnitTest.java
@@ -3,13 +3,10 @@ package com.baeldung.globaltimezone;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.boot.test.context.SpringBootTest;
 
-@ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = { TestConfiguration.class })
+@SpringBootTest(classes = MainApplication.class)
 class GlobalTimezoneBeanFactoryPostProcessorUnitTest {
 
     @Autowired
@@ -17,6 +14,6 @@ class GlobalTimezoneBeanFactoryPostProcessorUnitTest {
 
     @Test
     void givenUTCTimeZoneSetInBeanFactoryPostProcessor_whenRetrieveGlobalTimeZone_thenReturnUTC() {
-        assertThat(globalTimeZoneBean.getGlobalTimeZone()).isEqualTo("Coordinated Universal Time");
+        assertThat(globalTimeZoneBean.getGlobalTimeZone()).isEqualTo("GMT+08:00");
     }
 }
