--- spring-core-4/src/test/java/com/baeldung/globaltimezone/TestConfiguration.java
@@ -1,18 +1,0 @@
-package com.baeldung.globaltimezone;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class TestConfiguration {
-
-    @Bean
-    public GlobalTimezoneBeanFactoryPostProcessor gimezoneBeanFactoryPostProcessor() {
-        return new GlobalTimezoneBeanFactoryPostProcessor();
-    }
-
-    @Bean
-    public GlobalTimeZoneBean globalTimeZoneBean() {
-        return new GlobalTimeZoneBean();
-    }
-}
